requirements.txt

streamlit
pandas
matplotlib
seaborn
numpy
scipy
openpyxl
python-docx
